Group arguments into files by category
