+++ wrapper/minecraft/src/main/kotlin/org/quiltmc/qkl/wrapper/minecraft/brigadier/arguments/PrimitiveArguments.kt
@@ -1,0 +1,133 @@
+/*
+ * Copyright 2022 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Preserve binary compatibility when moving extensions between files
+ */
+@file:JvmMultifileClass
+@file:JvmName("ArgumentsKt")
+
+package org.quiltmc.qkl.wrapper.minecraft.brigadier.arguments
+
+import com.mojang.brigadier.arguments.*
+import com.mojang.brigadier.builder.ArgumentBuilder
+import com.mojang.brigadier.builder.RequiredArgumentBuilder
+import org.quiltmc.qkl.wrapper.minecraft.brigadier.RequiredArgumentAction
+
+/**
+ * Adds a boolean argument with [name] as the parameter name.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.boolean(
+    name: String,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, Boolean>(
+        name,
+        BoolArgumentType.bool()
+    )
+    argument.apply(action)
+    then(argument)
+}
+
+/**
+ * Adds a double argument with [name] as the parameter name.
+ *
+ * @param min the minimum value.
+ * @param max the maximum value.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.double(
+    name: String,
+    min: Double = -Double.MAX_VALUE,
+    max: Double = Double.MAX_VALUE,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, Double>(
+        name,
+        DoubleArgumentType.doubleArg(min, max)
+    )
+    argument.apply(action)
+    then(argument)
+}
+
+/**
+ * Adds a float argument with [name] as the parameter name.
+ *
+ * @param min the minimum value.
+ * @param max the maximum value.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.float(
+    name: String,
+    min: Float = -Float.MAX_VALUE,
+    max: Float = Float.MAX_VALUE,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, Float>(
+        name,
+        FloatArgumentType.floatArg(min, max)
+    )
+    argument.apply(action)
+    then(argument)
+}
+
+/**
+ * Adds an integer argument with [name] as the parameter name.
+ *
+ * @param min the minimum value.
+ * @param max the maximum value.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.integer(
+    name: String,
+    min: Int = -Int.MAX_VALUE,
+    max: Int = Int.MAX_VALUE,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, Int>(
+        name,
+        IntegerArgumentType.integer(min, max)
+    )
+    argument.apply(action)
+    then(argument)
+}
+
+/**
+ * Adds a long argument with [name] as the parameter name.
+ *
+ * @param min the minimum value.
+ * @param max the maximum value.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.long(
+    name: String,
+    min: Long = -Long.MAX_VALUE,
+    max: Long = Long.MAX_VALUE,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, Long>(
+        name,
+        LongArgumentType.longArg(min, max)
+    )
+    argument.apply(action)
+    then(argument)
+}
