+++ wrapper/minecraft/src/main/kotlin/org/quiltmc/qkl/wrapper/minecraft/brigadier/arguments/RangeArguments.kt
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2022 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Preserve binary compatibility when moving extensions between files
+ */
+@file:JvmMultifileClass
+@file:JvmName("ArgumentsKt")
+
+package org.quiltmc.qkl.wrapper.minecraft.brigadier.arguments
+
+import com.mojang.brigadier.builder.ArgumentBuilder
+import com.mojang.brigadier.builder.RequiredArgumentBuilder
+import net.minecraft.command.argument.NumberRangeArgumentType
+import net.minecraft.predicate.NumberRange
+import org.quiltmc.qkl.wrapper.minecraft.brigadier.RequiredArgumentAction
+
+/**
+ * Adds a float range argument with [name] as the parameter name.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.floatRange(
+    name: String,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, NumberRange.FloatRange>(
+        name,
+        NumberRangeArgumentType.floatRange()
+    )
+    argument.apply(action)
+    then(argument)
+}
+
+/**
+ * Adds an int range argument with [name] as the parameter name.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.intRange(
+    name: String,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, NumberRange.IntRange>(
+        name,
+        NumberRangeArgumentType.intRange()
+    )
+    argument.apply(action)
+    then(argument)
+}
