+++ wrapper/minecraft/src/main/kotlin/org/quiltmc/qkl/wrapper/minecraft/brigadier/arguments/TextArguments.kt
@@ -1,0 +1,83 @@
+/*
+ * Copyright 2022 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Preserve binary compatibility when moving extensions between files
+ */
+@file:JvmMultifileClass
+@file:JvmName("ArgumentsKt")
+
+package org.quiltmc.qkl.wrapper.minecraft.brigadier.arguments
+
+import com.mojang.brigadier.builder.ArgumentBuilder
+import com.mojang.brigadier.builder.RequiredArgumentBuilder
+import net.minecraft.command.argument.ColorArgumentType
+import net.minecraft.command.argument.MessageArgumentType
+import net.minecraft.command.argument.TextArgumentType
+import net.minecraft.text.Text
+import net.minecraft.util.Formatting
+import org.quiltmc.qkl.wrapper.minecraft.brigadier.RequiredArgumentAction
+
+/**
+ * Adds a message argument with [name] as the parameter name.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.message(
+    name: String,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, MessageArgumentType.MessageFormat>(
+        name,
+        MessageArgumentType()
+    )
+    argument.apply(action)
+    then(argument)
+}
+
+/**
+ * Adds a color argument with [name] as the parameter name.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.color(
+    name: String,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, Formatting>(
+        name,
+        ColorArgumentType.color()
+    )
+    argument.apply(action)
+    then(argument)
+}
+
+/**
+ * Adds a text argument with [name] as the parameter name.
+ *
+ * @author Oliver-makes-code (Emma)
+ */
+public fun <S> ArgumentBuilder<S, *>.text(
+    name: String,
+    action: RequiredArgumentAction<S>
+) {
+    val argument = RequiredArgumentBuilder.argument<S, Text>(
+        name,
+        TextArgumentType.text()
+    )
+    argument.apply(action)
+    then(argument)
+}
